It works a little better now

+++ MerkleHellmanKnapsack/Attacker.h
@@ -12,7 +12,11 @@ using namespace std;
 class Attacker
 {
 public:
+
+	///<summary></summary>
 	Attacker();
+
+	///<summary></summary>
 	~Attacker();
 
 	string BreakMerkleHellmanKnapsack(vector<unsigned long long> pubKey, vector<string> cipherText);
@@ -20,17 +24,17 @@ public:
 private:
 	int originalN = 0;
 	int originalM = 0;
-	vector<vector<float>> a;
+	vector<vector<double>> a;
 
-	vector<vector<float>> LLLReduce(vector<vector<float>> matrix);
-	vector<vector<float>> DoGramSchmidt(vector<vector<float>> matrix);
+	vector<vector<double>> LLLReduce(vector<vector<double>> matrix);
+	vector<vector<double>> DoGramSchmidt(vector<vector<double>> matrix);
 
-	float DotProduct(vector<float> a, vector<float> b);
-	vector<float> MultiplyVector(vector<float> a, float b);
-	vector<float> SubtractVector(vector<float> a, vector<float> b);
-	vector<float> AddVector(vector<float> a, vector<float> b);
-	bool IsReducedBasis(vector<vector<float>> matrix, vector<vector<float>> matrix1, vector<vector<float>> currentA);
-	bool CheckForZeros(vector<float> v);
-	vector<float> GetSolution(vector<vector<float>> matrix, vector<vector<float>> matrix1);
+	double DotProduct(vector<double> a, vector<double> b);
+	vector<double> MultiplyVector(vector<double> a, double b);
+	vector<double> SubtractVector(vector<double> a, vector<double> b);
+	vector<double> AddVector(vector<double> a, vector<double> b);
+	bool IsReducedBasis(vector<vector<double>> matrix, vector<vector<double>> matrix1, vector<vector<double>> currentA);
+	bool CheckForZeros(vector<double> v);
+	vector<double> GetSolution(vector<vector<double>> matrix, vector<vector<double>> matrix1);
 };
